image_compressor/requirements.txt
Pillow>=10.0.0
customtkinter>=5.2.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
tkinter-tooltip>=2.0.0
